+++ src/types/Block.ts
@@ -8,10 +8,15 @@ export type Block = {
 	// for developers of this library - the "isSensitive" attribute is technically processed only when user sends an
 	// input, but for ease of understanding to users, we will classify it with pre-processing attributes
 	message?: string | void | ((params: Params) => string | void) | ((params: Params) => Promise<string | void>);
-	options?: Array<string> | ((params: Params) => Array<string>) | ((params: Params) => Promise<Array<string>>);
-	checkboxes?: {items: Array<string>, max?: number, min?: number} |
-		((params: Params) => {items: Array<string>, max?: number, min?: number}) |
-		((params: Params) => Promise<{items: Array<string>, max?: number, min?: number}>);
+	options?: Array<string> | ((params: Params) => Array<string>) | ((params: Params) => Promise<Array<string>>) |
+		{items: Array<string>, sendOutput?: boolean, reusable?: boolean} |
+		((params: Params) => {items: Array<string>, sendOutput?: boolean, reusable?: boolean}) |
+		((params: Params) => Promise<{items: Array<string>, sendOutput?: boolean, reusable?: boolean}>);
+	checkboxes?: Array<string> | ((params: Params) => Array<string>) |
+		((params: Params) => Promise<Array<string>>) |
+		{items: Array<string>, max?: number, min?: number, sendOutput?: boolean, reusable?: boolean} |
+		((params: Params) => {items: Array<string>, max?: number, min?: number, sendOutput?: boolean, reusable?: boolean}) |
+		((params: Params) => Promise<{items: Array<string>, max?: number, min?: number, sendOutput?: boolean, reusable?: boolean}>);
 	component?: JSX.Element | void | ((params: Params) => JSX.Element | void) |
 	((params: Params) => Promise<JSX.Element | void>);
 	chatDisabled?: boolean | ((params: Params) => boolean) | ((params: Params) => Promise<boolean>);