+++ src/types/internal/Toast.ts
@@ -1,0 +1,8 @@
+/**
+ * Defines the available attributes for a toast.
+ */
+export type Toast = {
+    id: string;
+    content: string | JSX.Element;
+    timeout?: number;
+}