+++ src/utils/buttonBuilder.tsx
@@ -138,13 +138,16 @@ export const createFileAttachmentButton = (
 	getCurrPath: () => keyof Flow | null,
 	getPrevPath: () => keyof Flow | null,
 	goToPath: (pathToGo: keyof Flow) => void,
+	setTextAreaValue: (value: string) => void,
+	injectToast: (content: string | JSX.Element, timeout?: number) => void,
 	handleActionInput: (path: keyof Flow, userInput: string, sendUserInput?: boolean) => Promise<void>
 ) => {
 	return (
 		<FileAttachmentButton inputRef={inputRef} flow={flow} 
 			blockAllowsAttachment={blockAllowsAttachment} getCurrPath={getCurrPath} openChat={openChat}
 			getPrevPath={getPrevPath} goToPath={goToPath} handleActionInput={handleActionInput}
-			injectMessage={injectMessage} streamMessage={streamMessage}
+			injectMessage={injectMessage} streamMessage={streamMessage} setTextAreaValue={setTextAreaValue}
+			injectToast={injectToast}
 		/>
 	);
 };